[FEAT] Portal updates
