+++ src/features/portal/example/lib/PortalProvider.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useEffect } from "react";
 import { useInterpret } from "@xstate/react";
 import { MachineInterpreter, portalMachine } from "./portalMachine";
 
@@ -15,7 +15,27 @@ export const PortalProvider: React.FC = ({ children }) => {
     portalMachine
   ) as unknown as MachineInterpreter;
 
-  console.log({ provided: portalService });
+  /**
+   * Below is how we can listen to messages from the parent window
+   */
+  useEffect(() => {
+    const handleMessage = (event: MessageEvent) => {
+      // Handle the received message
+      if (event.data.event === "purchased") {
+        // Put in your handlers here
+        portalService.send("PURCHASED");
+      }
+    };
+
+    // Add event listener to listen for messages from the parent window
+    window.addEventListener("message", handleMessage);
+
+    // Clean up the event listener when component unmounts
+    return () => {
+      window.removeEventListener("message", handleMessage);
+    };
+  }, []);
+
   return (
     <PortalContext.Provider value={{ portalService }}>
       {children}
